# prod/Dragon3/backend/server.js
/**
 * ====================================================================
 * DRAGON3 - SERVIDOR CON EMBASSY (VERSIÓN PRODUCCIÓN)
 * ====================================================================
 *
 * Este servidor Express actúa como punto de entrada HTTP para el frontend.
 * Delega todo el análisis de imágenes al Agent Embassy (puerto 3002),
 * que gestiona colas, orquestador, sistema de defensa, evolución y adaptador FAANG.
 *
 * Funcionalidades:
 * - Autenticación JWT para rutas privadas.
 * - Subida de archivos con Multer (validación MIME, límite 100MB).
 * - Selección dinámica del plan según el tipo de archivo (mimetype).
 * - Proxy hacia /agent/execute del Embassy.
 * - Health check y frontend estático.
 * - Trazabilidad con correlationId y archivoId.
 *
 * Dependencias:
 * - Express, Multer, JWT, MongoDB (opcional), UUID, node-fetch.
 *
 * @module server
 * @version 2.0.0
 */

import dotenv from 'dotenv';
import path from 'path';
import { fileURLToPath } from 'url';
import express from 'express';
import cors from 'cors';
import helmet from 'helmet';
import compression from 'compression';
import multer from 'multer';
import jwt from 'jsonwebtoken';
import mongoose from 'mongoose';
import { v4 as uuidv4 } from 'uuid';
import fs from 'fs';
import fetch from 'node-fetch';
import client from 'prom-client';

// ============================================================
// 1. CONFIGURACIÓN INICIAL
// ============================================================

const __filename = fileURLToPath(import.meta.url);
const __dirname = path.dirname(__filename);

// Cargar variables de entorno (usando .env en producción)
dotenv.config({ path: path.resolve(__dirname, './.env') });

const MODULE_NAME = 'server-embassy';
const VERSION = '2.0.0';
const PORT = process.env.PORT || 3000;
const JWT_SECRET = process.env.JWT_SECRET || 'mi-secreto-temporal-123';
const EMBASSY_URL = process.env.EMBASSY_URL || 'http://localhost:3002';
const EMBASSY_EXECUTE_ENDPOINT = `${EMBASSY_URL}/agent/execute`;
const MAX_SYNC_ANALYSES = Math.max(1, Number.parseInt(process.env.MAX_SYNC_ANALYSES || '4', 10));
const EMBASSY_TIMEOUT_MS = Math.max(1000, Number.parseInt(process.env.EMBASSY_TIMEOUT_MS || '180000', 10));
const METRICS_ENABLED = process.env.ENABLE_METRICS !== 'false';
const METRICS_TOKEN = process.env.METRICS_TOKEN || '';
let activeAnalyses = 0;

const metricsRegistry = new client.Registry();
client.collectDefaultMetrics({ register: metricsRegistry, prefix: 'dragon3_' });
const httpRequestsTotal = new client.Counter({
  name: 'dragon3_http_requests_total',
  help: 'Total de peticiones HTTP recibidas por el backend',
  labelNames: ['method', 'route', 'status_code'],
  registers: [metricsRegistry]
});
const httpRequestDuration = new client.Histogram({
  name: 'dragon3_http_request_duration_seconds',
  help: 'Duracion de peticiones HTTP del backend',
  labelNames: ['method', 'route', 'status_code'],
  buckets: [0.05, 0.1, 0.25, 0.5, 1, 2, 5, 10, 30, 60, 180],
  registers: [metricsRegistry]
});
const activeAnalysisGauge = new client.Gauge({
  name: 'dragon3_active_analyses',
  help: 'Analisis sincronicos activos en esta instancia del backend',
  registers: [metricsRegistry]
});
const analysisCapacityRejections = new client.Counter({
  name: 'dragon3_analysis_capacity_rejections_total',
  help: 'Analisis rechazados por falta de slots sincronicos',
  registers: [metricsRegistry]
});

console.log(`🔧 [${MODULE_NAME}] Variables de entorno cargadas`);
console.log(`🤖 Embassy URL: ${EMBASSY_URL}`);

// ============================================================
// 2. CONEXIÓN A MONGODB (OPCIONAL)
// ============================================================
// Solo se usa para mantener sesión o guardar análisis, pero no es crítico.
let mongoConnection = null;

async function conectarMongoDB() {
  try {
    await mongoose.connect(process.env.MONGO_URI, {
      serverSelectionTimeoutMS: 5000,
    });
    mongoConnection = mongoose.connection;
    console.log(`✅ [${MODULE_NAME}] MongoDB conectado`);
  } catch (error) {
    console.error(`❌ [${MODULE_NAME}] Error conectando MongoDB:`, error.message);
    // No detenemos el servidor si MongoDB falla (opcional)
  }
}

// ============================================================
// 3. EXPRESS - CONFIGURACIÓN
// ============================================================

const app = express();

// Middlewares de seguridad y rendimiento
app.use(helmet({ contentSecurityPolicy: false }));
app.use(cors({ origin: true, credentials: true }));
app.use(compression());
app.use(express.json({ limit: '50mb' }));
app.use(express.urlencoded({ extended: true, limit: '50mb' }));

// Correlation ID para trazabilidad
app.use((req, res, next) => {
  req.correlationId = req.headers['x-correlation-id'] || uuidv4();
  req.startTime = Date.now();
  res.setHeader('X-Correlation-ID', req.correlationId);
  next();
});

app.use((req, res, next) => {
  const startedAt = process.hrtime.bigint();
  res.once('finish', () => {
    const route = req.route?.path || req.path || 'unknown';
    const labels = { method: req.method, route, status_code: String(res.statusCode) };
    const durationSeconds = Number(process.hrtime.bigint() - startedAt) / 1e9;
    httpRequestsTotal.inc(labels);
    httpRequestDuration.observe(labels, durationSeconds);
  });
  next();
});

// ============================================================
// 4. MIDDLEWARE DE AUTENTICACIÓN JWT
// ============================================================

/**
 * Middleware que verifica el token JWT en el header Authorization.
 * Si es válido, añade `req.usuario` con los datos decodificados.
 * Si no, devuelve 401 o 403.
 */
function authMiddleware(req, res, next) {
  const authHeader = req.headers.authorization;
  if (!authHeader || !authHeader.startsWith('Bearer ')) {
    return res.status(401).json({ error: 'Token requerido', correlationId: req.correlationId });
  }
  const token = authHeader.split(' ')[1];
  try {
    const decoded = jwt.verify(token, JWT_SECRET);
    req.usuario = decoded;
    next();
  } catch (error) {
    return res.status(403).json({ error: 'Token inválido', correlationId: req.correlationId });
  }
}

// ============================================================
// 5. MULTER - CONFIGURACIÓN DE SUBIDA
// ============================================================

const uploadPath = path.resolve(__dirname, 'uploads/temporal');
if (!fs.existsSync(uploadPath)) {
  fs.mkdirSync(uploadPath, { recursive: true });
}

const storage = multer.diskStorage({
  destination: (req, file, cb) => cb(null, uploadPath),
  filename: (req, file, cb) => {
    const ext = path.extname(file.originalname);
    const base = path.basename(file.originalname, ext);
    cb(null, `${base}-${Date.now()}-${uuidv4().slice(0, 8)}${ext}`);
  }
});

const fileFilter = (req, file, cb) => {
  const allowed = [
    'image/jpeg', 'image/png', 'image/webp',
    'application/pdf', 'video/mp4'
  ];
  if (allowed.includes(file.mimetype)) {
    cb(null, true);
  } else {
    cb(new Error('Tipo de archivo no permitido'), false);
  }
};

const upload = multer({
  storage,
  fileFilter,
  limits: { fileSize: 100 * 1024 * 1024 } // 100 MB
});

// ============================================================
// 6. FUNCIÓN PARA LLAMAR AL EMBASSY
// ============================================================

/**
 * Lee el archivo subido, lo convierte a base64, genera un token JWT
 * para el Embassy y envía la petición a /agent/execute.
 *
 * @param {string} filePath - Ruta del archivo temporal.
 * @param {string} nombreOriginal - Nombre original del archivo.
 * @param {string} correlationId - ID de correlación.
 * @param {string} archivoId - ID único del análisis.
 * @param {string|null} usuarioId - ID del usuario (si autenticado).
 * @param {string} planId - ID del plan a ejecutar (por defecto 'analizar-imagen-completa').
 * @returns {Promise<Object>} Respuesta del Embassy en formato FAANG.
 */
async function llamarEmbassy(
  filePath,
  nombreOriginal,
  correlationId,
  archivoId,
  usuarioId = null,
  planId = 'analizar-imagen-completa'
) {
  // Leer archivo y convertirlo a base64
  const fileBuffer = fs.readFileSync(filePath);
  const base64 = fileBuffer.toString('base64');

  // Generar token JWT para el Embassy (usamos el mismo secret, pero podríamos usar otro)
  const token = jwt.sign(
    {
      agentId: 'dragon3-internal',
      nivel: 'confianza' // Puede ser 'publico', 'habitual', 'confianza'
    },
    JWT_SECRET,
    { expiresIn: '1h' }
  );

  // Construir payload según lo que espera el Embassy
  const payload = {
    token,
    agentId: 'dragon3-internal',
    peticion: {
      archivo: base64,
      params: {}, // Parámetros adicionales si se necesitan
      extra: {
        archivoId,
        nombreOriginal,
        usuarioId,
        correlationId
      }
    },
    configuracion: {
      planId, // Usar el plan que ya está corregido con rutas absolutas
      // También podríamos usar tipoArchivo para auto-detección:
      // tipoArchivo: 'imagen'
    }
  };

  console.log(`📡 [${MODULE_NAME}] Llamando al Embassy con plan: ${planId}`);

  const response = await fetch(EMBASSY_EXECUTE_ENDPOINT, {
    method: 'POST',
    headers: {
      'Content-Type': 'application/json',
      'X-Correlation-ID': correlationId
    },
    body: JSON.stringify(payload),
    signal: AbortSignal.timeout(EMBASSY_TIMEOUT_MS)
  });

  if (!response.ok) {
    const errorText = await response.text();
    throw new Error(`Embassy error ${response.status}: ${errorText}`);
  }

  // El Embassy ya devuelve el formato FAANG completo (con resultado, imagenId, etc.)
  return await response.json();
}

function crearTokenEmbassy(agentId, ownerId) {
  return jwt.sign({ agentId, ownerId, nivel: 'confianza' }, JWT_SECRET, { expiresIn: '10m' });
}

async function llamarEmbassyJobs(endpoint, payload, correlationId) {
  const ownerId = payload.ownerId;
  const response = await fetch(`${EMBASSY_URL}${endpoint}`, {
    method: 'POST',
    headers: {
      'Content-Type': 'application/json',
      'X-Correlation-ID': correlationId,
      Authorization: `Bearer ${crearTokenEmbassy('dragon3-async-api', ownerId)}`
    },
    body: JSON.stringify({ ...payload, token: crearTokenEmbassy('dragon3-async-api', ownerId) }),
    signal: AbortSignal.timeout(15000)
  });
  const body = await response.json().catch(() => ({}));
  if (!response.ok) {
    throw new Error(body.error || `Embassy jobs error ${response.status}`);
  }
  return body;
}

async function consultarEmbassyJob(jobId, ownerId, correlationId) {
  const response = await fetch(`${EMBASSY_URL}/agent/analysis-jobs/${encodeURIComponent(jobId)}`, {
    headers: {
      'X-Correlation-ID': correlationId,
      'Content-Type': 'application/json',
      Authorization: `Bearer ${crearTokenEmbassy('dragon3-async-api', ownerId)}`
    },
    signal: AbortSignal.timeout(10000)
  });
  const body = await response.json().catch(() => ({}));
  if (!response.ok) throw new Error(body.error || `Embassy jobs error ${response.status}`);
  return body;
}

async function cancelarEmbassyJob(jobId, ownerId, correlationId) {
  const response = await fetch(`${EMBASSY_URL}/agent/analysis-jobs/${encodeURIComponent(jobId)}`, {
    method: 'DELETE',
    headers: {
      'X-Correlation-ID': correlationId,
      Authorization: `Bearer ${crearTokenEmbassy('dragon3-async-api', ownerId)}`
    },
    signal: AbortSignal.timeout(10000)
  });
  const body = await response.json().catch(() => ({}));
  if (!response.ok) throw new Error(body.error || `Embassy jobs error ${response.status}`);
  return body;
}

function tryAcquireAnalysisSlot() {
  if (activeAnalyses >= MAX_SYNC_ANALYSES) {
    return null;
  }

  activeAnalyses += 1;
  activeAnalysisGauge.inc();
  let released = false;
  return () => {
    if (!released) {
      released = true;
      activeAnalyses -= 1;
      activeAnalysisGauge.dec();
    }
  };
}

function rejectWhenAtCapacity(req, res) {
  const releaseSlot = tryAcquireAnalysisSlot();
  if (!releaseSlot) {
    analysisCapacityRejections.inc();
    res.set('Retry-After', '5');
    res.status(429).json({
      error: 'Capacidad de análisis temporalmente agotada',
      correlationId: req.correlationId
    });
    return null;
  }

  return releaseSlot;
}
// ============================================================
// 7. RUTAS
// ============================================================

// 7.1 Health check
app.get('/health', (req, res) => {
  res.json({
    status: 'ok',
    timestamp: new Date().toISOString(),
    version: VERSION,
    services: {
      mongodb: mongoConnection?.readyState === 1 ? 'connected' : 'disconnected',
      embassy: EMBASSY_URL
    }
  });
});

app.get('/metrics', async (req, res) => {
  if (!METRICS_ENABLED) {
    return res.status(404).end();
  }

  const providedToken = req.get('X-Metrics-Token') || '';
  const remoteAddress = req.socket.remoteAddress || '';
  const localRequest = remoteAddress === '127.0.0.1' ||
    remoteAddress === '::1' ||
    remoteAddress === '::ffff:127.0.0.1';
  if (METRICS_TOKEN ? providedToken !== METRICS_TOKEN : !localRequest) {
    return res.status(404).end();
  }

  res.set('Content-Type', metricsRegistry.contentType);
  res.end(await metricsRegistry.metrics());
});

/**
 * Determina el plan a ejecutar según el tipo MIME del archivo subido.
 * @param {string} mimetype - Tipo MIME del archivo.
 * @returns {string} ID del plan.
 */
function seleccionarPlanPorMimetype(mimetype) {
  if (mimetype === 'image/jpeg' || mimetype === 'image/jpg') {
    return 'analizar-imagen-completa';   // Plan con ML (XGBoost)
  } else if (mimetype === 'image/png') {
    return 'analizar-imagen-sin-ml';     // Plan sin ML (para PNG)
  } else {
    // Para otros formatos (webp, pdf, etc.) usamos el plan sin ML por defecto
    return 'analizar-imagen-sin-ml';
  }
}

// 7.2 Análisis público (sin autenticación)
app.post('/analizar-imagen-publico', upload.single('archivo'), async (req, res) => {
  const correlationId = req.correlationId;
  const archivoId = uuidv4();
  const releaseSlot = rejectWhenAtCapacity(req, res);

  if (!releaseSlot) {
    if (req.file?.path) {
      try { fs.unlinkSync(req.file.path); } catch (error) { /* ignorar */ }
    }
    return;
  }

  try {
    if (!req.file) {
      return res.status(400).json({ error: 'No se ha subido ningún archivo', correlationId });
    }

    // Seleccionar plan según el tipo MIME
    const planId = seleccionarPlanPorMimetype(req.file.mimetype);

    console.log(`📤 [${MODULE_NAME}] Análisis público iniciado: ${req.file.originalname} (${correlationId}) - Plan: ${planId}`);

    const resultado = await llamarEmbassy(
      req.file.path,
      req.file.originalname,
      correlationId,
      archivoId,
      null, // usuarioId
      planId
    );

    // El resultado ya es el FAANG completo
    res.json(resultado);
  } catch (error) {
    console.error(`❌ [${MODULE_NAME}] Error en análisis público:`, error.message);
    if (req.file?.path) {
      try { fs.unlinkSync(req.file.path); } catch (e) {}
    }
    res.status(500).json({
      error: 'Error en el servidor',
      correlationId,
      details: process.env.NODE_ENV === 'development' ? error.message : undefined
    });
  } finally {
    releaseSlot();
    if (req.file?.path) {
      try { fs.unlinkSync(req.file.path); } catch (error) { /* ignorar */ }
    }
  }
});

// 7.3 Análisis autenticado
app.post('/api/analizar-imagen', authMiddleware, upload.single('archivo'), async (req, res) => {
  const correlationId = req.correlationId;
  const archivoId = uuidv4();
  const usuarioId = req.usuario?.id || req.usuario?.userId || null;
  const releaseSlot = rejectWhenAtCapacity(req, res);

  if (!releaseSlot) {
    if (req.file?.path) {
      try { fs.unlinkSync(req.file.path); } catch (error) { /* ignorar */ }
    }
    return;
  }

  try {
    if (!req.file) {
      return res.status(400).json({ error: 'No se ha subido ningún archivo', correlationId });
    }

    // Seleccionar plan según el tipo MIME
    const planId = seleccionarPlanPorMimetype(req.file.mimetype);

    console.log(`📤 [${MODULE_NAME}] Análisis autenticado: ${req.file.originalname} (${correlationId}, usuario: ${usuarioId}) - Plan: ${planId}`);

    const resultado = await llamarEmbassy(
      req.file.path,
      req.file.originalname,
      correlationId,
      archivoId,
      usuarioId,
      planId
    );

    // Aquí podrías guardar en MongoDB si lo necesitas
    // const analisis = new AnalisisArchivo({ ... });
    // await analisis.save();

    res.json(resultado);
  } catch (error) {
    console.error(`❌ [${MODULE_NAME}] Error en análisis autenticado:`, error.message);
    if (req.file?.path) {
      try { fs.unlinkSync(req.file.path); } catch (e) {}
    }
    res.status(500).json({
      error: 'Error en el servidor',
      correlationId,
      details: process.env.NODE_ENV === 'development' ? error.message : undefined
    });
  } finally {
    releaseSlot();
    if (req.file?.path) {
      try { fs.unlinkSync(req.file.path); } catch (error) { /* ignorar */ }
    }
  }
});

// 7.4 Frontend estático (opcional)
const frontendPath = path.resolve(__dirname, '../frontend');
if (fs.existsSync(frontendPath)) {
  app.use(express.static(frontendPath));
  app.get('/', (req, res) => {
    res.sendFile(path.join(frontendPath, 'index.html'));
  });
  app.get('/analizar-imagen-publico.html', (req, res) => {
    res.sendFile(path.join(frontendPath, 'analizar-imagen-publico.html'));
  });
  console.log(`🌐 [${MODULE_NAME}] Frontend servido desde: ${frontendPath}`);
} else {
  console.warn(`⚠️ [${MODULE_NAME}] Frontend no encontrado en: ${frontendPath}`);
}

app.post('/api/v1/analysis-jobs', authMiddleware, upload.single('archivo'), async (req, res) => {
  const correlationId = req.correlationId;
  const ownerId = String(req.usuario?.id || req.usuario?.userId || req.usuario?.sub || 'unknown');
  if (!req.file) {
    return res.status(400).json({ error: 'No se ha subido ningún archivo', correlationId });
  }

  try {
    const planId = seleccionarPlanPorMimetype(req.file.mimetype);
    const resultado = await llamarEmbassyJobs('/agent/analysis-jobs', {
      ownerId,
      filePath: req.file.path,
      nombreOriginal: req.file.originalname,
      archivoId: uuidv4(),
      correlationId,
      planId,
      idempotencyKey: req.get('Idempotency-Key') || null
    }, correlationId);
    res.status(202).json(resultado);
  } catch (error) {
    try { fs.unlinkSync(req.file.path); } catch (cleanupError) { /* ignorar */ }
    res.status(502).json({ error: 'No se pudo encolar el análisis', correlationId });
  }
});

app.get('/api/v1/analysis-jobs/:jobId', authMiddleware, async (req, res) => {
  const ownerId = String(req.usuario?.id || req.usuario?.userId || req.usuario?.sub || 'unknown');
  try {
    res.json(await consultarEmbassyJob(req.params.jobId, ownerId, req.correlationId));
  } catch (error) {
    res.status(404).json({ error: 'Trabajo no encontrado', correlationId: req.correlationId });
  }
});

app.delete('/api/v1/analysis-jobs/:jobId', authMiddleware, async (req, res) => {
  const ownerId = String(req.usuario?.id || req.usuario?.userId || req.usuario?.sub || 'unknown');
  try {
    res.status(202).json(await cancelarEmbassyJob(req.params.jobId, ownerId, req.correlationId));
  } catch (error) {
    res.status(404).json({ error: 'Trabajo no encontrado', correlationId: req.correlationId });
  }
});

// ============================================================
// 8. INICIAR SERVIDOR
// ============================================================

async function startServer() {
  // Conectar MongoDB (opcional, no crítico)
  await conectarMongoDB();

  // Iniciar Express
  const server = app.listen(PORT, '127.0.0.1', () => {
    console.log(`🚀 [${MODULE_NAME}] Servidor Dragon3 (con Embassy) en http://localhost:${PORT}`);
    console.log(`📚 Health: http://localhost:${PORT}/health`);
    console.log(`🔍 Análisis público: http://localhost:${PORT}/analizar-imagen-publico`);
    console.log(`🔒 Análisis autenticado: http://localhost:${PORT}/api/analizar-imagen`);
    console.log(`🤖 Embassy en: ${EMBASSY_URL}`);
  });

  // Graceful shutdown
  const gracefulShutdown = async (signal) => {
    console.log(`🛑 [${MODULE_NAME}] Recibido ${signal}, cerrando...`);
    server.close(() => {
      console.log('✅ Servidor HTTP cerrado');
    });
    if (mongoConnection) {
      await mongoose.disconnect();
      console.log('✅ MongoDB desconectado');
    }
    process.exit(0);
  };

  process.on('SIGTERM', () => gracefulShutdown('SIGTERM'));
  process.on('SIGINT', () => gracefulShutdown('SIGINT'));

  return server;
}

// ============================================================
// 9. EJECUCIÓN
// ============================================================

startServer().catch(error => {
  console.error(`❌ [${MODULE_NAME}] Error al iniciar:`, error);
  process.exit(1);
});

// ============================================================
// 10. EXPORTACIÓN (por si se necesita en pruebas)
// ============================================================

export default app;